@@ -7,14 +7,14 @@
     "http-proxy-middleware": "0.19.1",
     "moment": "^2.24.0",
     "node-sass": "^4.13.1",
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
+    "react": "16.13.0",
+    "react-dom": "16.13.0",
     "react-scripts": "2.1.8"
   },
   "scripts": {
     "start": "cross-env HTTPS=true react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "jest",
     "eject": "react-scripts eject"
   },
   "browserslist": [
@@ -24,6 +24,12 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "cross-env": "6.0.3"
+    "@babel/preset-env": "^7.8.6",
+    "@babel/preset-react": "^7.8.3",
+    "@testing-library/react": "^9.4.1",
+    "babel-jest": "^25.1.0",
+    "cross-env": "6.0.3",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^25.1.0"
   }
 }
