@@ -1,10 +1,10 @@
 // Copyright (C) 2007-2019, GoodData(R) Corporation. All rights reserved.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { render } from '@testing-library/react'
 import App from './App';
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  const {container} = render(<App />)
+  expect(container).toBeDefined();
 });
